# api/src/HelloBear.Api/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "HelloBear API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Community/{id}/{type}": {
      "put": {
        "tags": [
          "Community"
        ],
        "operationId": "Community_ReactCommunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReactionType"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Community/{id}": {
      "delete": {
        "tags": [
          "Community"
        ],
        "operationId": "Community_DeleteCommunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Community/upload-image": {
      "post": {
        "tags": [
          "Community"
        ],
        "operationId": "Community_UploadImage",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "StudentId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "ClassId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "LessionId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Image": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  },
                  "Thumbnail": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Community/upload-video": {
      "post": {
        "tags": [
          "Community"
        ],
        "operationId": "Community_UploadVideo",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "StudentId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "ClassId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "LessionId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Title": {
                    "type": "string",
                    "nullable": true
                  },
                  "Video": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/Community/media-status": {
      "get": {
        "tags": [
          "Community"
        ],
        "operationId": "Community_GetMediaStatus",
        "parameters": [
          {
            "name": "RequestId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMediaStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/TextBook/{Id}": {
      "get": {
        "tags": [
          "TextBook"
        ],
        "operationId": "TextBook_GetTextbookDetail",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextBookDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/TextBook/unit": {
      "get": {
        "tags": [
          "TextBook"
        ],
        "operationId": "TextBook_GetUnitsWithPagination",
        "parameters": [
          {
            "name": "TextbookId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "SearchText",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "SortCriteria",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "$ref": "#/components/schemas/SortCriteria"
              }
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListOfUnitResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/TextBook/upload-thumbnail": {
      "post": {
        "tags": [
          "TextBook"
        ],
        "operationId": "TextBook_UploadThumbnail",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/TextBook/{id}": {
      "delete": {
        "tags": [
          "TextBook"
        ],
        "operationId": "TextBook_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/TextBook/all": {
      "get": {
        "tags": [
          "TextBook"
        ],
        "operationId": "TextBook_GetAllTextBooks",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextBookQueryResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/User/{Id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetUserDetail",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/User/roles": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetRoles",
        "parameters": [
          {
            "name": "SearchText",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "SortCriteria",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "$ref": "#/components/schemas/SortCriteria"
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListOfRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/User/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "User_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "bodyRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserBodyRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "User_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/User/{id}/resend-invitation": {
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "User_SendInvitation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/User/all-teachers": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetAllTeachers",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/GetAllTeachersQuery"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAllTeachersResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SortCriteria": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "sortKey": {
            "type": "string"
          },
          "isDescending": {
            "type": "boolean"
          }
        }
      },
...
      "ReactionType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Like",
          "Heart"
        ],
        "enum": [
          1,
          2
        ]
      },
      "GetMediaStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StudentMediaStatus"
          },
          "bytesSent": {
            "type": "integer",
            "format": "int64"
          },
          "exception": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StudentMediaStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "NotStarted",
          "Starting",
          "Uploading",
          "Completed",
          "Failed"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4
        ]
      },
...
      "TextBookQueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "hasClasses": {
            "type": "boolean"
          }
        }
      },
      "TextBookDetailResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PaginatedListOfUnitResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
      "UnitResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UserStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Invited",
          "Actived",
          "Deactivated"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "UserDetailResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PhoneType"
              }
            ]
          }
        }
      },
      "PhoneType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Landline",
          "Mobile"
        ],
        "enum": [
          1,
          2
        ]
      },
      "PaginatedListOfRoleResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
      "RoleResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateUserBodyRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PhoneType"
              }
            ]
          }
        }
      },
      "GetAllTeachersResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
      "GetAllTeachersQuery": {
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}